[TASK] Replace Fluid Widgets
Replace Pagination Widget with georgringer/numbered_pagination
Replace Map Widget with Controller Action supposed to return JavaScript (does not work yet)
