+++ .editorconfig
@@ -18,7 +18,7 @@ indent_size = 4
 # HTML-Files
 [*.html]
 indent_style = space
-indent_size = 2
+indent_size = 4
 
 # TMPL-Files
 [*.tmpl]
