+++ Resources/Private/Partials/Pagination.html
@@ -1,0 +1,59 @@
+<nav aria-label="Seiten blättern">
+  <ul class="pagination">
+      <f:if condition="{pagination.previousPageNumber} && {pagination.previousPageNumber} >= {pagination.firstPageNumber}">
+          <f:then>
+              <f:comment>
+                  <li class="first">
+                      <a href="{f:uri.action(action:actionName, arguments:{currentPage: 1})}" title="{f:translate(key:'pagination.first')}">
+                          <i class="material-icons">first_page</i>
+                      </a>
+                  </li>
+              </f:comment>
+              <li class="previous">
+                  <a href="{f:uri.action(action:actionName, arguments:{currentPage: pagination.previousPageNumber})}" aria-label="{f:translate(key:'pagination.previous')}">
+                    <span aria-hidden="true">«</span>
+                  </a>
+              </li>
+          </f:then>
+          <f:else>
+              <f:comment>
+                  <li class="disabled"><span><i class="material-icons">first_page</i></span></li>
+                  <li class="disabled"><span>{f:translate(key:'widget.pagination.previous', extensionName: 'fluid')}</span></li>
+              </f:comment>
+          </f:else>
+      </f:if>
+      <f:if condition="{pagination.hasLessPages}">
+          <li>…</li>
+      </f:if>
+      <f:for each="{pagination.allPageNumbers}" as="page">
+          <li class="{f:if(condition: '{page} == {paginator.currentPageNumber}', then:'current')}">
+              <a href="{f:uri.action(action:actionName, arguments:{currentPage: page})}">{page}</a>
+          </li>
+      </f:for>
+      <f:if condition="{pagination.hasMorePages}">
+          <li>…</li>
+      </f:if>
+      <f:if condition="{pagination.nextPageNumber} && {pagination.nextPageNumber} <= {pagination.lastPageNumber}">
+          <f:then>
+              <li class="next">
+                  <a href="{f:uri.action(action:actionName, arguments:{currentPage: pagination.nextPageNumber})}" aria-label="{f:translate(key:'pagination.next')}">
+                     <span aria-hidden="true">»</span>
+                  </a>
+              </li>
+              <f:comment>
+                  <li class="last">
+                      <a href="{f:uri.action(action:actionName, arguments:{currentPage: pagination.lastPageNumber})}" title="{f:translate(key:'pagination.last')}">
+                          last
+                      </a>
+                  </li>
+              </f:comment>
+          </f:then>
+          <f:else>
+              <f:comment>
+                  <li class="disabled"><span>{f:translate(key:'widget.pagination.next', extensionName: 'fluid')}</span></li>
+                  <li class="disabled"><span>last page</span></li>
+              </f:comment>
+          </f:else>
+      </f:if>
+  </ul>
+</nav>
