+++ Resources/Private/Templates/Entry/List.html
@@ -9,24 +9,24 @@
     <f:if condition="{entries}">
         <f:then>
             <div class="list-group">
-                <bzga:widget.paginate as="paginatedEntries" demand="{demand}"
-                                      configuration="{itemsPerPage: settings.list.itemsPerPage, insertAbove: 1, insertBelow: 1, maximumNumberOfLinks: 10}"
-                                      objects="{entries}">
-                    <f:for each="{paginatedEntries}" as="entry">
-                        <f:link.action pageUid="{settings.singlePid}" controller="Entry" action="show"
-                                       arguments="{entry: entry}" class="list-group-item">
-                            <h4 class="list-group-item-heading">{entry.title} <f:render partial="Entry/Distance" arguments="{_all}" /></h4>
-                            <div class="list-group-item-text">
-                                {entry.address}
-                            </div>
-                        </f:link.action>
-                    </f:for>
-                </bzga:widget.paginate>
+                <f:render partial="Pagination" arguments="{pagination: pagination.pagination, paginator: pagination.paginator, actionName: 'list'}" />
+                <f:for each="{pagination.paginator.paginatedItems}" as="entry" iteration="iterator">
+                    <f:link.action pageUid="{settings.singlePid}" controller="Entry" action="show"
+                                   arguments="{entry: entry}" class="list-group-item">
+                        <h4 class="list-group-item-heading">{entry.title}
+                            <f:render partial="Entry/Distance" arguments="{_all}"/>
+                        </h4>
+                        <div class="list-group-item-text">
+                            {entry.address}
+                        </div>
+                    </f:link.action>
+                </f:for>
+                <f:render partial="Pagination" arguments="{pagination: pagination.pagination, paginator: pagination.paginator, actionName: 'list'}" />
             </div>
         </f:then>
         <f:else>
             <div class="alert alert-info" role="alert">
-                <f:translate key="no-entries-found"></f:translate>
+                <f:translate key="no-entries-found" />
             </div>
         </f:else>
     </f:if>
