+++ Resources/Private/Templates/Entry/Show.html
@@ -41,5 +41,5 @@
         <bzga:implode glue=", ">{entry.categories}</bzga:implode>
         <hr />
     </f:if>
-    <bzga:widget.map demand="{demand}" entry="{entry}" settings="{settings}"/>
+    <div style="width: 100%; height: 300px; " id="{mapId}"></div>
 </f:section>
