+++ composer.json
@@ -40,7 +40,7 @@
         "php-http/guzzle7-adapter": "^1.0",
         "php-http/message": "^1.7",
         "geocoder-php/nominatim-provider": "^5.1",
-        "phpstan/phpstan": "^1.9.14"
+        "georgringer/numbered-pagination": "^1.0"
     },
     "require-dev": {
         "typo3/testing-framework": "^6.14.0",
@@ -57,7 +57,8 @@
         "rector/rector": "^0.12.13",
         "phpspec/prophecy": "^1.16",
         "nimut/testing-framework": "^6.0",
-        "phpspec/prophecy-phpunit": "^2.0"
+        "phpspec/prophecy-phpunit": "^2.0",
+        "phpstan/phpstan": "^1.9.14"
     },
     "replace": {
         "typo3-ter/bzga-beratungsstellensuche": "self.version"
