+++ phpstan-baseline.neon
@@ -11,16 +11,6 @@ parameters:
 			path: Classes/Domain/Map/Leaflet/MapBuilder.php
 
 		-
-			message: "#^Property Bzga\\\\BzgaBeratungsstellensuche\\\\Domain\\\\Model\\\\Dto\\\\Demand\\:\\:\\$categories \\(iterable\\<Bzga\\\\BzgaBeratungsstellensuche\\\\Domain\\\\Model\\\\Category\\>&TYPO3\\\\CMS\\\\Extbase\\\\Persistence\\\\ObjectStorage\\) does not accept TYPO3\\\\CMS\\\\Extbase\\\\Persistence\\\\ObjectStorage\\.$#"
-			count: 2
-			path: Classes/Domain/Model/Dto/Demand.php
-
-		-
-			message: "#^Property Bzga\\\\BzgaBeratungsstellensuche\\\\Domain\\\\Model\\\\Entry\\:\\:\\$categories \\(iterable\\<Bzga\\\\BzgaBeratungsstellensuche\\\\Domain\\\\Model\\\\Category\\>&TYPO3\\\\CMS\\\\Extbase\\\\Persistence\\\\ObjectStorage\\) does not accept TYPO3\\\\CMS\\\\Extbase\\\\Persistence\\\\ObjectStorage\\.$#"
-			count: 2
-			path: Classes/Domain/Model/Entry.php
-
-		-
 			message: "#^Call to an undefined method TYPO3\\\\CMS\\\\Extbase\\\\Persistence\\\\QueryInterface\\:\\:statement\\(\\)\\.$#"
 			count: 1
 			path: Classes/Domain/Repository/EntryRepository.php
@@ -44,23 +34,3 @@ parameters:
 			message: "#^Property Bzga\\\\BzgaBeratungsstellensuche\\\\Service\\\\SettingsService\\:\\:\\$settings \\(array\\) on left side of \\?\\? is not nullable\\.$#"
 			count: 1
 			path: Classes/Service/SettingsService.php
-
-		-
-			message: "#^Call to an undefined method TYPO3\\\\CMS\\\\Extbase\\\\DomainObject\\\\DomainObjectInterface\\:\\:getInfoWindow\\(\\)\\.$#"
-			count: 1
-			path: Classes/ViewHelpers/Widget/Controller/MapController.php
-
-		-
-			message: "#^Call to an undefined method TYPO3\\\\CMS\\\\Extbase\\\\DomainObject\\\\DomainObjectInterface\\:\\:getLatitude\\(\\)\\.$#"
-			count: 1
-			path: Classes/ViewHelpers/Widget/Controller/MapController.php
-
-		-
-			message: "#^Call to an undefined method TYPO3\\\\CMS\\\\Extbase\\\\DomainObject\\\\DomainObjectInterface\\:\\:getLongitude\\(\\)\\.$#"
-			count: 1
-			path: Classes/ViewHelpers/Widget/Controller/MapController.php
-
-		-
-			message: "#^Method Bzga\\\\BzgaBeratungsstellensuche\\\\ViewHelpers\\\\Widget\\\\Controller\\\\MapController\\:\\:getTyposcriptFrontendController\\(\\) is unused\\.$#"
-			count: 1
-			path: Classes/ViewHelpers/Widget/Controller/MapController.php
